+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # @biconomy/sdk
 
+## 0.0.21
+
+### Patch Changes
+
+- Add support for token paymaster with helper functions
+
+## 0.0.20
+
+### Patch Changes
+
+- Counterfactual address helper export
+
 ## 0.0.18
 
 ### Patch Changes